Fix message item costs and next message ID misalignment (#96)

+++ mm/src/code/z_message.c
@@ -2160,13 +2160,13 @@ void Message_SetupLoadItemIcon(PlayState* play) {
         msgCtx->nextTextId = font->msgBuf.schar[++msgCtx->msgBufPos] << 8;
         msgCtx->nextTextId |= font->msgBuf.schar[++msgCtx->msgBufPos];
 
-        msgCtx->unk1206C = font->msgBuf.schar[++msgCtx->msgBufPos] << 8;
+        msgCtx->unk1206C = (u8)(font->msgBuf.schar[++msgCtx->msgBufPos] << 8);
         msgCtx->unk1206C |= font->msgBuf.schar[++msgCtx->msgBufPos];
 
-        msgCtx->unk12070 = font->msgBuf.schar[++msgCtx->msgBufPos] << 8;
+        msgCtx->unk12070 = (u8)(font->msgBuf.schar[++msgCtx->msgBufPos] << 8);
         msgCtx->unk12070 |= font->msgBuf.schar[++msgCtx->msgBufPos];
 
-        msgCtx->unk12074 = font->msgBuf.schar[++msgCtx->msgBufPos] << 8;
+        msgCtx->unk12074 = (u8)(font->msgBuf.schar[++msgCtx->msgBufPos] << 8);
         msgCtx->unk12074 |= font->msgBuf.schar[++msgCtx->msgBufPos];
 
         msgCtx->msgBufPos++;
